activities refactored, added layouts for navigation drawer
