+++ app/build.gradle
@@ -21,6 +21,8 @@ android {
 dependencies {
     implementation fileTree(include: ['*.jar'], dir: 'libs')
     implementation 'androidx.appcompat:appcompat:1.1.0'
+    implementation "androidx.navigation:navigation-fragment:2.1.0"
+    implementation "androidx.navigation:navigation-ui:2.1.0"
     implementation 'androidx.legacy:legacy-support-v4:1.0.0'
     testImplementation 'junit:junit:4.12'
     androidTestImplementation 'androidx.test:runner:1.2.0'
@@ -32,11 +34,11 @@ dependencies {
     api 'com.squareup.okhttp3:logging-interceptor:4.2.1'
     api 'androidx.annotation:annotation:1.1.0'
     implementation project(path: ':svgloader')
-    implementation 'com.google.android.material:material:1.0.0'
-    implementation 'com.google.firebase:firebase-analytics:17.2.0'
-    implementation 'com.google.firebase:firebase-messaging:20.0.0'
-    implementation 'com.google.firebase:firebase-core:17.2.0'
-    implementation 'com.google.firebase:firebase-auth:19.1.0'
+    api 'com.google.android.material:material:1.0.0'
+    api 'com.google.firebase:firebase-analytics:17.2.0'
+    api 'com.google.firebase:firebase-messaging:20.0.0'
+    api 'com.google.firebase:firebase-core:17.2.0'
+    api 'com.google.firebase:firebase-auth:19.1.0'
 }
 
 apply plugin: 'com.google.gms.google-services'
