+++ app/src/main/AndroidManifest.xml
@@ -3,6 +3,10 @@
     xmlns:tools="http://schemas.android.com/tools"
     package="com.example.footballapi">
 
+    <uses-permission android:name="android.permission.INTERNET" />
+
+    <uses-permission android:name="android.permission.ACCESS_NETWORK_STATE" />
+
     <application
         android:icon="@mipmap/ic_logo"
         android:label="@string/app_name"
@@ -33,14 +37,11 @@
         <activity android:name=".view.activities.ConnexionActivity" />
         <activity android:name=".view.activities.InscriptionActivity" />
 
-        <service
-            android:name=".service.FootLiteFirebaseMessagingService"
-            android:enabled="true"
-            android:exported="true" />
-
+        <service android:name=".FootLiteFirebaseMessagingService">
+            <intent-filter>
+                <action android:name="com.google.firebase.MESSAGING_EVENT"/>
+            </intent-filter>
+        </service>
     </application>
-    <uses-permission android:name="android.permission.ACCESS_NETWORK_STATE" />
-
-    <uses-permission android:name="android.permission.INTERNET" />
 
 </manifest>
