+++ app/src/main/java/com/example/footballapi/view/activities/MainActivity.java
@@ -32,7 +32,7 @@ public class MainActivity extends AppCompatActivity implements View.OnClickListe
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.main_activity);
+        setContentView(R.layout.activity_main);
 
         // On change le title de l'actionBar
         this.setTitle("Compétitions");
