+++ app/src/main/res/layout/activity_competition.xml
