+++ app/src/main/res/layout/activity_inscription.xml
