+++ app/src/main/res/layout/activity_player.xml
